延迟计算，时间校准

+++ CSharp/NewRuntime/Net/LatencyResult.cs
@@ -4,6 +4,6 @@ namespace UselessFrame.Net
     public struct LatencyResult
     {
         public bool Success;
-        public float DeltaTime;
+        public int DeltaTime;
     }
 }
